Part and Preamble related code is extracted to the separate files.
